ROFO-24 refactor: profileImage을 common으로 분리
